# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "TrustPay Test Api",
        "version": "1.0.0"
    },
    "basePath": "/apitest",
    "paths": {
...
        "/customer/forgot": {
            "post": {
                "tags": [
                    "Buyer Module"
                ],
                "summary": "Customer Forgot Password",
                "parameters": [
                    {
                        "name": "customer",
                        "in": "body",
                        "description": "Customer email to get email with reset token",
                        "required": true,
                        "schema": {
                            "properties": {
                                "email": {
                                    "description": "Email Address",
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{status: success, message: token_generated, data: [token: reset_token_key ]}"
                    },
                    "201": {
                        "description": "{status: error, message: user_not_found}"
                    }
                }
            }
        },
        "/seller/forgot": {
            "post": {
                "tags": [
                    "Seller Module"
                ],
                "summary": "Seller Forgot Password",
                "parameters": [
                    {
                        "name": "seller",
                        "in": "body",
                        "description": "Seller email to get email with reset token",
                        "required": true,
                        "schema": {
                            "properties": {
                                "email": {
                                    "description": "Email Address",
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{status: success, message: token_generated, data: [token: reset_token_key ]}"
                    },
                    "201": {
                        "description": "{status: error, message: user_not_found}"
                    }
                }
            }
        },
...
        "/customer/forgot/reset": {
            "post": {
                "tags": [
                    "Buyer Module"
                ],
                "summary": "Customer Reset Forgot Password",
                "parameters": [
                    {
                        "name": "customer",
                        "in": "body",
                        "description": "Customer email, new password and reset token to reset password",
                        "required": true,
                        "schema": {
                            "properties": {
                                "email": {
                                    "description": "Email Address",
                                    "type": "string"
                                },
                                "password": {
                                    "description": "New Password",
                                    "type": "string"
                                },
                                "password_confirmation": {
                                    "description": "Confirm New Password",
                                    "type": "string"
                                },
                                "token": {
                                    "description": "Reset Token from email/forgot response",
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{status: success, message: reset_successful}"
                    },
                    "201": {
                        "description": "{status: error, message: reset_failed}"
                    }
                }
            }
        },
        "/seller/forgot/reset": {
            "post": {
                "tags": [
                    "Seller Module"
                ],
                "summary": "Seller Reset Forgot Password",
                "parameters": [
                    {
                        "name": "seller",
                        "in": "body",
                        "description": "Seller email, new password and reset token to reset password",
                        "required": true,
                        "schema": {
                            "properties": {
                                "email": {
                                    "description": "Email Address",
                                    "type": "string"
                                },
                                "password": {
                                    "description": "New Password",
                                    "type": "string"
                                },
                                "password_confirmation": {
                                    "description": "Confirm New Password",
                                    "type": "string"
                                },
                                "token": {
                                    "description": "Reset Token from email/forgot response",
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{status: success, message: reset_successful}"
                    },
                    "201": {
                        "description": "{status: error, message: reset_failed"
                    }
                }
            }
        },
...
    },
...
}